Added support for websocket timeouts
