+++ config/websocket_proxy_settings.conf
@@ -1,4 +1,4 @@
 include proxy_settings.conf;
-proxy_connect_timeout 7d;
-proxy_read_timeout 7d;
-proxy_send_timeout 7d;
+proxy_connect_timeout 4h;
+proxy_read_timeout 4h;
+proxy_send_timeout 4h;
